feat: run worktree setup in a setup terminal instead of blocking workspace creation
Creating a linked workspace held the New Workspace dialog open until every
`worktree.setup` command finished, which is why the RPC carried a 30 minute
timeout. The desktop now asks the host to prepare the setup instead: the dialog
closes as soon as the Git worktree exists, and the workspace opens with its
usual Terminal 1 plus a Setup terminal where the work is visible.

The Setup terminal runs one portable line against a host-generated script.
`&&` cannot be used because the terminal hosts whatever interactive shell the
user configured and PowerShell 5.1 rejects it at parse time; writing one command
per line up front cannot be used either, because PTY bytes reach the foreground
process, so the second line would land on the first command's stdin. The script
sequences instead, echoing each command and continuing past a failure, and it
deletes itself when it finishes. The copy rules go through a new
`alera workspace setup --copies-only` so their symlink and path-escape
validation stays in Rust.

`deferSetup`, `deferredSetupCommand` and the `initialCommandOnce` tab payload
key are additive: a host that does not understand them runs the setup inline and
omits the command, which is the previous behavior.
