+++ AGENTS.md
@@ -175,6 +175,7 @@ When planning is needed, use a spec-driven development flow. Do not jump straigh
 - Recurring host work that nobody reads while nobody is asking MUST go through `DemandDrivenTicker` (`rust/alera-cli/src/terminal_host/demand_driven_ticker.rs`) rather than an ad hoc `tokio::spawn` loop. The host is a sidecar and cannot see whether the app's window is visible, and it MUST NOT take a client's word for it either, because a client that reports going away may instead have died mid-report. Silence is the signal that survives both, so the ticker stops on an idle window and restarts on the next request. Pick an idle window comfortably longer than the client's polling period, or a client polling on time still finds the ticker stopped.
 - A client that fell behind is resynchronised from the delivery cursor the host keeps per client (`Session::delivered_output_cursors`), never by resending the scrollback. It falls behind two ways, a visibility pause and a full output queue, and both take the same path through `resume_output_for_client` (`rust/alera-cli/src/terminal_host/server/output_delivery.rs`). The cursor MUST advance only when a frame is *accepted* by that client's queue: advancing it on append makes a dropped frame a silent hole. The missed bytes MUST go out on the terminal lane ahead of the unpause, not inside the request's reply, because the reply travels on the control lane and the writer drains terminal frames first, so bytes returned inline can land after output that came later; the lane also feeds them through the client's per-session UTF-8 decoder, so a code point split at the pause boundary still joins up. A client the host cannot place in the stream gets a full snapshot instead, which is the only correct answer once the ring has dropped the gap. The Dart client MUST NOT discard output between losing visibility and the pause taking effect, because the host already counted those frames as delivered.
 - The snapshot an attach or a resync replays is capped by `restoreSnapshotBytes`, which is separate from `scrollbackBytes` on purpose: the first is what a client's emulator will keep, the second is what the host retains so `terminal.read` and the coordinator tail can page back through it. Both are additive and MUST NOT bump `aleraTerminalHostProtocolVersion`; a host that receives no cap replays the whole buffer, which is what an older app expects.
+- The desktop defers a project's worktree setup to a terminal tab named `Setup` instead of holding the New Workspace dialog open until `pnpm install` finishes. `deferSetup` on `workspace.createManaged`, `deferredSetupCommand` on its response, and the `initialCommandOnce` tab payload key are additive and MUST NOT bump `aleraTerminalHostProtocolVersion`; a host that ignores the flag runs the setup inline and omits the command, which is exactly the old behavior. The tab runs one portable line (`/bin/sh "<script>"`, `cmd /d /c "<script>"`) against a host-generated script, and that indirection is load-bearing. `&&` cannot be used: the terminal hosts whatever interactive shell the user configured, PowerShell 5.1 rejects `&&` at parse time and nushell removed it. Writing one command per line up front cannot be used either, because PTY bytes go to the *foreground process*, so the second line lands on the first command's stdin. The Windows launcher omits `/s` on purpose - with it, cmd strips the outer quotes and a script path containing spaces breaks apart - and leaves `cmd` unquoted so PowerShell does not need the `&` call operator. Copy rules go through `alera workspace setup --copies-only` rather than being rewritten in shell, so `copy_rule_inner`'s symlink and path-escape validation stays in Rust. `initialCommandOnce` exists because agent tabs deliberately re-mint their `initialCommand` on every new PTY, so the clearing MUST stay opt-in.
 - Runtime change events carry an optional scope id (`workspaceId`, `projectId`) and an absent or empty scope means wildcard: every watcher refreshes. The app can attach to an already-running sidecar, so an older host broadcasting an empty payload MUST keep working. Emit these events through the helpers in `rust/alera-cli/src/terminal_host/server/runtime_change_broadcasts.rs` and pass `None` whenever the mutation really is broader than one workspace or project. Never guess a scope: the wildcard is the safe value, a wrong id silently stops watchers from updating. Adding a scope field is additive and MUST NOT bump `aleraTerminalHostProtocolVersion`, because a version mismatch makes the app treat a live host as unusable.
 
 ## Computer Use
