+++ rust/alera-cli/src/cli.rs
@@ -7,11 +7,13 @@ mod browser;
 mod computer;
 mod emulator;
 mod mobile;
+mod workspace;
 
 pub use browser::*;
 pub use computer::*;
 pub use emulator::*;
 pub use mobile::*;
+pub use workspace::*;
 
 use clap::{Args, Parser, Subcommand, ValueEnum};
 /// Top-level CLI, mirroring the Dart `AleraCliCommandRunner`.
@@ -297,142 +299,6 @@ pub struct ProjectAddArgs {
 pub enum ProjectKindArg {
     GitRepository,
     Folder,
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceCommand {
-    #[command(flatten)]
-    pub runtime: RuntimeDirArgs,
-    #[command(flatten)]
-    pub output: OutputArgs,
-    #[command(subcommand)]
-    pub action: WorkspaceAction,
-}
-
-#[derive(Debug, Subcommand)]
-pub enum WorkspaceAction {
-    /// List workspaces for one project or all projects.
-    List(WorkspaceListArgs),
-    /// Create an Alera-managed Git worktree workspace.
-    Add(WorkspaceAddArgs),
-    /// Remove an Alera-managed Git worktree workspace.
-    Remove(WorkspaceRemoveArgs),
-    /// Register a workspace record without touching Git worktrees.
-    Register(WorkspaceRegisterArgs),
-    /// Remove a workspace record and related runtime records without touching Git worktrees.
-    Unregister(IdArgs),
-    /// Pin a workspace in the desktop sidebar.
-    Pin(IdArgs),
-    /// Unpin a workspace from the desktop sidebar.
-    Unpin(IdArgs),
-    /// Add a parent/child relationship.
-    Link(WorkspaceLinkArgs),
-    /// Remove a parent/child relationship.
-    Unlink(WorkspaceLinkArgs),
-    /// Assign a tag to a workspace.
-    Tag(WorkspaceTagArgs),
-    /// Remove a tag from a workspace.
-    Untag(WorkspaceTagArgs),
-    /// Preview opt-in cascade targets.
-    CascadePreview(CascadePreviewArgs),
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceListArgs {
-    #[arg(long = "project-id")]
-    pub project_id: Option<String>,
-    #[arg(long)]
-    pub all: bool,
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceAddArgs {
-    #[arg(long)]
-    pub id: Option<String>,
-    #[arg(long = "project-id")]
-    pub project_id: String,
-    #[arg(long)]
-    pub branch: String,
-    #[arg(long = "source-branch")]
-    pub source_branch: Option<String>,
-    #[arg(long)]
-    pub name: Option<String>,
-    #[arg(long = "reuse-existing-branch")]
-    pub reuse_existing_branch: bool,
-    #[arg(long = "workspace-root", conflicts_with = "path")]
-    pub workspace_root: Option<String>,
-    #[arg(long, conflicts_with = "workspace_root")]
-    pub path: Option<String>,
-    #[arg(long = "parent-workspace-id")]
-    pub parent_workspace_id: Option<String>,
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceRemoveArgs {
-    #[arg(long)]
-    pub id: String,
-    #[arg(long = "delete-branch", conflicts_with = "keep_branch")]
-    pub delete_branch: bool,
-    #[arg(long = "keep-branch", conflicts_with = "delete_branch")]
-    pub keep_branch: bool,
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceRegisterArgs {
-    #[arg(long)]
-    pub id: Option<String>,
-    #[arg(long = "instance-id")]
-    pub instance_id: Option<String>,
-    #[arg(long = "host-id")]
-    pub host_id: Option<String>,
-    #[arg(long = "project-id")]
-    pub project_id: String,
-    #[arg(long)]
-    pub name: String,
-    #[arg(long)]
-    pub path: String,
-    #[arg(long)]
-    pub branch: Option<String>,
-    #[arg(long = "source-branch")]
-    pub source_branch: Option<String>,
-    #[arg(long, value_enum, default_value_t = WorkspaceKindArg::Linked)]
-    pub kind: WorkspaceKindArg,
-    #[arg(long = "reuse-existing-branch")]
-    pub reuses_existing_branch: bool,
-}
-
-#[derive(Debug, Clone, Copy, ValueEnum)]
-pub enum WorkspaceKindArg {
-    Main,
-    Linked,
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceLinkArgs {
-    #[arg(long = "parent-workspace-id")]
-    pub parent_workspace_id: String,
-    #[arg(long = "child-workspace-id")]
-    pub child_workspace_id: String,
-}
-
-#[derive(Debug, Args)]
-pub struct WorkspaceTagArgs {
-    #[arg(long = "workspace-id")]
-    pub workspace_id: String,
-    #[arg(long = "tag-id")]
-    pub tag_id: String,
-}
-
-#[derive(Debug, Args)]
-pub struct CascadePreviewArgs {
-    #[arg(long = "workspace-id")]
-    pub workspace_ids: Vec<String>,
-    #[arg(long = "tag-id")]
-    pub tag_ids: Vec<String>,
-    #[arg(long = "descendants")]
-    pub include_descendants: bool,
-    #[arg(long = "tags")]
-    pub include_tags: bool,
 }
 
 #[derive(Debug, Args)]
