+++ rust/alera-cli/src/terminal_host/server/runtime_mutation_barrier.rs
@@ -18,6 +18,7 @@ pub(super) fn conflicts_with_runtime_mutation(request_type: &str) -> bool {
     matches!(
         request_type,
         "workspace.createManaged"
+            | "workspace.runSetup"
             | "createOrAttach"
             | "write"
             | "terminate"
